+++ FlexEngine/include/Helpers.hpp
@@ -30,6 +30,8 @@ namespace flex
 	extern ImVec4 g_WarningButtonHoveredColor;
 	extern ImVec4 g_WarningButtonActiveColor;
 
+	static const char* SEPARATOR_STR = ", ";
+
 	GLFWimage LoadGLFWimage(const std::string& filePath, i32 requestedChannelCount = 3, bool flipVertically = false, i32* channelCountOut = nullptr);
 	void DestroyGLFWimage(GLFWimage& image);
 
@@ -138,9 +140,9 @@ namespace flex
 	// String will be padded with '0's to be at least minChars long (excluding a leading '-' for negative numbers)
 	std::string IntToString(i32 i, u16 minChars = 0);
 
-	std::string Vec2ToString(const glm::vec2& vec, i32 precision);
-	std::string Vec3ToString(const glm::vec3& vec, i32 precision);
-	std::string Vec4ToString(const glm::vec4& vec, i32 precision);
+	std::string Vec2ToString(glm::vec2 vec, i32 precision);
+	std::string Vec3ToString(glm::vec3 vec, i32 precision);
+	std::string Vec4ToString(glm::vec4 vec, i32 precision);
 
 	void CopyVec3ToClipboard(const glm::vec3& vec);
 	void CopyVec4ToClipboard(const glm::vec4& vec);